+++ flutter_meta_wearables_dat_mock_device/ios/flutter_meta_wearables_dat_mock_device/Package.swift
@@ -3,7 +3,7 @@ import PackageDescription
 
 let package = Package(
     name: "flutter_meta_wearables_dat_mock_device",
-    platforms: [.iOS("17.0")],
+    platforms: [.iOS("17.2")],
     products: [
         .library(name: "flutter-meta-wearables-dat-mock-device", targets: ["flutter_meta_wearables_dat_mock_device"])
     ],